% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccf-package.R
\docType{package}
\name{rccf-package}
\alias{rccf}
\alias{rccf-package}
\title{rccf: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lmaowisc/rccf}
  \item \url{https://lmaowisc.github.io/rccf/}
  \item Report bugs at \url{https://github.com/lmaowisc/rccf/issues}
}

}
\keyword{internal}
