% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccf-package.R
\docType{package}
\name{rccf-package}
\alias{rccf}
\alias{rccf-package}
\title{rccf: Cumulative Frequency of Recurrent Events in Presence of Death}
\description{
Nonparametric estimation and testing of the cumulative incidence function of recurrent events in the presence of death (Ghosh and Lin, 2000, \doi{10.1111/j.0006-341x.2000.00554.x}). Also implements joint test with log-rank statistic on death.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lmaowisc/rccf}
  \item \url{https://lmaowisc.github.io/rccf/}
  \item Report bugs at \url{https://github.com/lmaowisc/rccf/issues}
}

}
\author{
\strong{Maintainer}: Lu Mao \email{lmao@biostat.wisc.edu} (\href{https://orcid.org/0000-0002-8626-9822}{ORCID})

}
\keyword{internal}
